+++ rust/src/locale.rs
@@ -770,7 +770,6 @@ locale_keys! {
     PanelAllProvidersShort,
     PanelShowAllProviders,
     PanelShowFewerProviders,
-    PanelZoom,
     PanelMenu,
     PanelCopied,
     PanelToday,